Move CLI to mcp
