+++ pyproject.toml
@@ -47,7 +47,7 @@ dev = [
 "Source" = "https://github.com/MinishLab/semble"
 
 [project.scripts]
-semble = "semble.cli:main"
+semble = "semble.mcp:main"
 
 [tool.setuptools]
 package-dir = {"" = "src"}
@@ -69,7 +69,6 @@ target-version = "py310"
 
 [tool.ruff.lint.per-file-ignores]
 "tests/**" = ["ANN"]
-"src/semble/cli.py" = ["T20"]
 "benchmarks/*.py" = ["T20"]
 
 [tool.ruff.lint]
